refactor(test): split controller_service tests.rs god-file into tests/ subdir
